requirements.txt
fastapi
httpx
pillow
pydantic
regex
requests
uvicorn
